+++ R/ranges.R
@@ -6,26 +6,37 @@ NULL
 #' 
 #' Get ranges (pairs of left and right corpus positions) for queries.
 #' 
-#' @param mc If \code{logical} value \code{TRUE}, the value of
-#'   \code{getOption("polmineR.cores")} is passed into \code{mclapply} or
-#'   \code{pblapply} as the specification of the number of cores to use. It is
-#'   also possible to supply an integer value with the number of cores directly.
-#'   Defaults to 1 (no multicore). Relevant only if several queries are to be
-#'   processed.
-#' @param progress A \code{logical} value, whether to show a progess bar when
+#' @param mc If `TRUE`, the value of `getOption("polmineR.cores")` is passed
+#'   into `mclapply()` or `pblapply()` as the specification of the number of
+#'   cores to use. It is also possible to supply an integer value with the
+#'   number of cores directly. Defaults to 1 (no multicore). Relevant only if
+#'   several queries are to be processed.
+#' @param progress A `logical` value, whether to show a progess bar when
 #'   processing multiple queries.
 #' @inheritParams cpos
 #' @rdname ranges
 #' @exportMethod ranges
+#' @examples
+#' use("RcppCWB")
+#' 
+#' dt <- corpus("REUTERS") %>% 
+#'   ranges(query = '"crude" "oil"', cqp = TRUE) %>% 
+#'   as.data.table()
 setGeneric("ranges", function(.Object, ...) standardGeneric("ranges"))
 
+#' @importFrom stringi stri_c_list
 #' @rdname ranges
 setMethod("ranges", "corpus", function(.Object, query, cqp = FALSE, check = TRUE, p_attribute = "word", verbose = TRUE, mc = 1L, progress = FALSE){
   if (is.logical(mc)) if (mc) mc <- getOption("polmineR.cores") else mc <- 1L
-  .fn <- function(x){
-    cpos(.Object = .Object, query = x, cqp = cqp, check = check, p_attribute = p_attribute, verbose = FALSE)
-  }
-  
+  .fn <- function(x) cpos(
+    .Object = .Object,
+    query = x,
+    cqp = cqp,
+    check = check,
+    p_attribute = p_attribute,
+    verbose = FALSE
+  )
+
   if (!is.null(names(query))) if (any(nchar(names(query)) == 0L)) stop(
     "If vector query is named, all individual queries need to be named."
   )
@@ -46,6 +57,14 @@ setMethod("ranges", "corpus", function(.Object, query, cqp = FALSE, check = TRUE
   y@query <- unlist(
     lapply(names(cpos_list), function(x) rep(x, times = nrow(cpos_list[[x]])))
   )
+  y@match <- stri_c_list(get_token_stream(
+    y@cpos,
+    p_attribute = p_attribute,
+    corpus = .Object@corpus,
+    registry = .Object@registry_dir,
+    split = TRUE
+  ), sep = " ")
+  
   y
 })
 
@@ -73,6 +92,14 @@ setMethod("ranges", "subcorpus", function(.Object, query, cqp = FALSE, check = T
   )
   rng@cpos <- rng@cpos[strucs_matches %in% .Object@strucs,]
   rng@query <- rng@query[strucs_matches %in% .Object@strucs]
+  rng@match <- stri_c_list(get_token_stream(
+    rng@cpos,
+    p_attribute = p_attribute,
+    corpus = .Object@corpus,
+    registry = .Object@registry_dir,
+    split = TRUE
+  ), sep = " ")
+  
   rng
 })
 
@@ -83,7 +110,7 @@ setMethod("ranges", "partition", function(.Object, query, cqp = FALSE, check = T
   ranges(.Object = as(.Object, "subcorpus"), query = query, cqp = cqp, check = check, p_attribute = p_attribute, verbose = verbose, mc = mc, progress = progress)
 })
 
-#' @param x A \code{ranges} class object.
+#' @param x A `ranges` class object.
 #' @param ... Additional arguments (unused).
 #' @rdname ranges_class
 #' @export
@@ -92,5 +119,6 @@ as.data.table.ranges <- function(x, ...){
   y <- data.table::as.data.table(x@cpos)
   colnames(y) <- c("cpos_left", "cpos_right")
   y[, "query" := x@query]
+  y[, "match" := x@match]
   y
 }